add sqitch for migrations
