+++ api/config.js
@@ -1,0 +1,28 @@
+const uid = require('uid-safe');
+const Auth0Strategy = require("passport-auth0");
+
+const sessionConfig = {
+  secret: uid.sync(18),
+  cookie: {
+    maxAge: 86400 * 1000 // 24 hours in milliseconds
+  },
+  resave: false,
+  saveUninitialized: true
+};
+
+const auth0Strategy = new Auth0Strategy(
+  {
+    domain: process.env.AUTH0_DOMAIN,
+    clientID: process.env.AUTH0_CLIENT_ID,
+    clientSecret: process.env.AUTH0_CLIENT_SECRET,
+    callbackURL: process.env.AUTH0_CALLBACK_URL
+  },
+  function(accessToken, refreshToken, extraParams, profile, done) {
+    return done(null, profile);
+  }
+);
+
+module.exports = { 
+  sessionConfig,
+  auth0Strategy
+};
