+++ api/helpers.js
@@ -1,0 +1,8 @@
+const restrictAccess = (req, res, next) => {
+  if (!req.isAuthenticated()) return res.redirect("/login");
+  next();
+};
+
+module.exports = {
+  restrictAccess,
+};
