+++ db/deploy/appschema.sql
@@ -1,0 +1,13 @@
+-- Deploy appschema
+BEGIN;
+
+CREATE SCHEMA tome;
+
+CREATE FUNCTION "update_changetimestamp_column" () RETURNS trigger LANGUAGE plpgsql AS '
+BEGIN
+   NEW.updated_at = now(); 
+   RETURN NEW;
+END;
+';
+
+COMMIT;
