+++ db/revert/appschema.sql
@@ -1,0 +1,7 @@
+-- Revert appschema
+BEGIN;
+
+DROP FUNCTION "update_changetimestamp_column";
+DROP SCHEMA tome;
+
+COMMIT;
