+++ db/sqitch.conf
@@ -1,0 +1,13 @@
+[core]
+	engine = pg
+	# plan_file = sqitch.plan
+	# top_dir = .
+	# deploy_dir = deploy
+	# revert_dir = revert
+	# verify_dir = verify
+	# extension = sql
+[user]
+	name = Daniel Sandoval
+	email = daniel@desandoval.net
+[engine "pg"]
+	registry = sqitch
