+++ db/sqitch.plan
@@ -1,0 +1,4 @@
+%syntax-version=1.0.0
+%project=tome
+
+appschema 2019-07-20T02:24:34Z Daniel Sandoval <daniel@desandoval.net> # Adds schema
