+++ db/verify/appschema.sql
@@ -1,0 +1,9 @@
+-- Verify anymessage
+
+BEGIN;
+
+SELECT pg_catalog.has_schema_privilege('tome', 'usage');
+
+select exists(select * from pg_proc where proname = 'update_changetimestamp_column');
+
+ROLLBACK;
