+++ server.js
@@ -1,70 +1,38 @@
-require("dotenv").config();
+const dev = process.env.NODE_ENV !== "production";
+if (dev) {
+  require("dotenv").config();
+}
+
 const express = require("express");
 const http = require("http");
 const next = require("next");
 const session = require("express-session");
-// 1 - importing dependencies
 const passport = require("passport");
-const Auth0Strategy = require("passport-auth0");
-const uid = require('uid-safe');
+
+const { sessionConfig, auth0Strategy } = require("./api/config");
+const { restrictAccess } = require("./api/helpers");
+
 const authRoutes = require("./api/auth-routes");
 const thoughtsAPI = require("./api/thoughts-api");
 
-const dev = process.env.NODE_ENV !== "production";
 const app = next({ dev });
 const handle = app.getRequestHandler();
 
 app.prepare().then(() => {
   const server = express();
 
-  // 2 - add session management to Express
-  const sessionConfig = {
-    secret: uid.sync(18),
-    cookie: {
-      maxAge: 86400 * 1000 // 24 hours in milliseconds
-    },
-    resave: false,
-    saveUninitialized: true
-  };
+  // auth config
   server.use(session(sessionConfig));
-
-  // 3 - configuring Auth0Strategy
-  const auth0Strategy = new Auth0Strategy(
-    {
-      domain: process.env.AUTH0_DOMAIN,
-      clientID: process.env.AUTH0_CLIENT_ID,
-      clientSecret: process.env.AUTH0_CLIENT_SECRET,
-      callbackURL: process.env.AUTH0_CALLBACK_URL
-    },
-    function(accessToken, refreshToken, extraParams, profile, done) {
-      return done(null, profile);
-    }
-  );
-
-  // 4 - configuring Passport
   passport.use(auth0Strategy);
   passport.serializeUser((user, done) => done(null, user));
   passport.deserializeUser((user, done) => done(null, user));
-
-  // 5 - adding Passport and authentication routes
   server.use(passport.initialize());
   server.use(passport.session());
+
+  // routes
   server.use(authRoutes);
-
-  // non-auth routes declared here
-
   server.use(thoughtsAPI);
-
-  // 6 - you are restricting access to some routes (redirect before render)
-  const restrictAccess = (req, res, next) => {
-    if (!req.isAuthenticated()) return res.redirect("/login");
-    next();
-  };
-
   server.use("/", restrictAccess);
-  // server.use("/share-thought", restrictAccess);
-
-  // handling everything else with Next.js
   server.get("*", handle);
 
   http.createServer(server).listen(process.env.PORT, () => {
